+++ pkgs/checks/CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fix a bug when using asynchronous conditions with `mayEmit` or
   `mayEmitMultiple`. Note that extensions using `nestAsync` should synchronously
   forward exceptions from that call.
+- Output more compact failure messages for simple expectations.
 - Fix a bug when printing a failure message involving a value that returns an
   empty output from `toString()`.